Correction des commandes rtorrent Préparation des "liens ticket" (sans authentification et c'est dernier unique) pour streaming Préparation de la partie Série Ajout du retour lors d'un ajout torrent

+++ webroot/bower_components/modernizr/feature-detects/workers-blobworkers.js
@@ -5,44 +5,62 @@
 
 // The test will potentially add garbage to console.
 
-(function () {
+(function(){
+  try {
+    // we're avoiding using Modernizr._domPrefixes as the prefix capitalization on
+    // these guys are notoriously peculiar.
+    var BlobBuilder = window.MozBlobBuilder || window.WebKitBlobBuilder || window.MSBlobBuilder || window.OBlobBuilder || window.BlobBuilder;
+    var URL         = window.MozURL || window.webkitURL || window.MSURL || window.OURL || window.URL;
+    var data    = 'Modernizr',
+        blob,
+        bb,
+        worker,
+        url,
+        timeout,
+        scriptText = 'this.onmessage=function(e){postMessage(e.data)}';
+
     try {
-
-        // we're avoiding using Modernizr._domPrefixes as the prefix capitalization on
-        // these guys are notoriously peculiar.
-        var BlobBuilder = window.MozBlobBuilder || window.WebKitBlobBuilder || window.MSBlobBuilder || window.OBlobBuilder || window.BlobBuilder,
-            URL = window.MozURL || window.webkitURL || window.MSURL || window.OURL || window.URL;
-
-        var data = 'Modernizr',
-            bb = new BlobBuilder();
-
-        bb.append('this.onmessage=function(e){postMessage(e.data)}');
-
-        var url = URL.createObjectURL(bb.getBlob()),
-            worker = new Worker(url);
-
-        bb = null;
-
-        worker.onmessage = function (e) {
-            worker.terminate();
-            URL.revokeObjectURL(url);
-            Modernizr.addTest('blobworkers', data === e.data);
-            worker = null;
-        };
-
-        // Just in case...
-        worker.onerror = function () {
-            Modernizr.addTest('blobworkers', false);
-            worker = null;
-        };
-
-        setTimeout(function () {
-            Modernizr.addTest('blobworkers', false);
-        }, 200);
-
-        worker.postMessage(data);
-
-    } catch (e) {
-        Modernizr.addTest('blobworkers', false);
+      blob = new Blob([scriptText], {type:'text/javascript'});
+    } catch(e) {
+      // we'll fall back to the deprecated BlobBuilder
     }
+    if (!blob) {
+      bb = new BlobBuilder();
+      bb.append(scriptText);
+      blob = bb.getBlob();
+    }
+
+    url = URL.createObjectURL(blob);
+    worker = new Worker(url);
+
+    worker.onmessage = function(e) {
+      Modernizr.addTest('blobworkers', data === e.data);
+      cleanup();
+    };
+
+    // Just in case...
+    worker.onerror = fail;
+    timeout = setTimeout(fail, 200);
+
+    worker.postMessage(data);
+  } catch (e) {
+    fail();
+  }
+
+  function fail() {
+    Modernizr.addTest('blobworkers', false);
+    cleanup();
+  }
+
+  function cleanup() {
+    if (url) {
+      URL.revokeObjectURL(url);
+    }
+    if (worker) {
+      worker.terminate();
+    }
+    if (timeout) {
+      clearTimeout(timeout);
+    }
+  }
 }());
